requirements.txt
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.31.0
flask==3.0.0
apscheduler==3.10.4
prometheus_client==0.19.0

# Test dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
